# Profile benchmark copy request response