--- a/handler_success.go
+++ b/handler_success.go
@@ -96,7 +96,7 @@ func estimateTagsCapacity(session *user.SessionState, apiSpec *APISpec) int {
 	return size
 }
 
-func (s *SuccessHandler) RecordHit(r *http.Request, timing int64, code int, requestCopy *http.Request, responseCopy *http.Response) {
+func (s *SuccessHandler) RecordHit(r *http.Request, timing int64, code int, responseCopy *http.Response) {
 
 	if s.Spec.DoNotTrack {
 		return
@@ -137,7 +137,7 @@ func (s *SuccessHandler) RecordHit(r *http.Request, timing int64, code int, requ
 		if recordDetail(r, s.Spec.GlobalConfig) {
 			// Get the wire format representation
 			var wireFormatReq bytes.Buffer
-			requestCopy.Write(&wireFormatReq)
+			r.Write(&wireFormatReq)
 			rawRequest = base64.StdEncoding.EncodeToString(wireFormatReq.Bytes())
 			// responseCopy, unlike requestCopy, can be nil
 			// here - if the response was cached in
@@ -264,11 +264,6 @@ func (s *SuccessHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) *http
 
 	addVersionHeader(w, r, s.Spec.GlobalConfig)
 
-	var copiedRequest *http.Request
-	if recordDetail(r, s.Spec.GlobalConfig) {
-		copiedRequest = copyRequest(r)
-	}
-
 	t1 := time.Now()
 	resp := s.Proxy.ServeHTTP(w, r)
 	t2 := time.Now()
@@ -277,12 +272,7 @@ func (s *SuccessHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) *http
 	log.Debug("Upstream request took (ms): ", millisec)
 
 	if resp != nil {
-		var copiedResponse *http.Response
-		if recordDetail(r, s.Spec.GlobalConfig) {
-			copiedResponse = copyResponse(resp)
-		}
-
-		s.RecordHit(r, int64(millisec), resp.StatusCode, copiedRequest, copiedResponse)
+		s.RecordHit(r, int64(millisec), resp.StatusCode, resp)
 	}
 	log.Debug("Done proxy")
 	return nil
@@ -309,27 +299,17 @@ func (s *SuccessHandler) ServeHTTPWithCache(w http.ResponseWriter, r *http.Reque
 		r.URL.RawPath = strings.Replace(r.URL.RawPath, s.Spec.Proxy.ListenPath, "", 1)
 	}
 
-	var copiedRequest *http.Request
-	if recordDetail(r, s.Spec.GlobalConfig) {
-		copiedRequest = copyRequest(r)
-	}
-
 	t1 := time.Now()
 	inRes := s.Proxy.ServeHTTPForCache(w, r)
 	t2 := time.Now()
 
 	addVersionHeader(w, r, s.Spec.GlobalConfig)
 
-	var copiedResponse *http.Response
-	if recordDetail(r, s.Spec.GlobalConfig) {
-		copiedResponse = copyResponse(inRes)
-	}
-
 	millisec := float64(t2.UnixNano()-t1.UnixNano()) * 0.000001
 	log.Debug("Upstream request took (ms): ", millisec)
 
 	if inRes != nil {
-		s.RecordHit(r, int64(millisec), inRes.StatusCode, copiedRequest, copiedResponse)
+		s.RecordHit(r, int64(millisec), inRes.StatusCode, inRes)
 	}
 
 	return inRes